app/services/admin-api-8031/requirements.txt
fastapi
uvicorn[standard]
asyncpg
pydantic
dnspython
httpx
email-validator
